requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
decorator==4.4.2
Django==2.1.5
django-cors-headers==3.0.2
django-cors-headers-multi==1.2.0
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
idna==2.10
IPy==1.0
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5